backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
mysqlclient==2.2.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
